Added old VS2008 project to mercurial.
In the original version, the source files were in a different directory than the
project files. I put the files in the project directory and updated the paths in
Real Matrices and Vectors.vcproj.
The file utility.h originated from the 'Real Matrices and Vectors' project but
was on a different directory (General Purpose). I put it in the project
directory together with the other source files and updated the paths in
Real Matrices and Vectors.vcproj and Real Matrices and Vectors.h.

Original timestamps for source files range from 26/8/2008 to 13/9/2008.
For project files:
Real Matrices and Vectors.sln:                              6/ 8/2008  5:21:21pm
Real Matrices and Vectors.vcproj:                          12/ 9/2008  6:14:19am
Real Matrices and Vectors.vcproj.SOULTAKER.Nightmare.user: 17/10/2008  3:45:23pm