Added the newer versions of 2 files.
The new tic-toc.cpp is not meant to replace the old and currently the headers do
not include corresponding prototypes.

Compared to the old tic-toc.cpp, the new uses C++ include names and namespaces.
The utility.h adds the sum, prod and identity template functions.
Also adds some comments.

This is an attempt to save all preserved versions of the files.

Original timespamps:
utility.h:              15/10/2008  2:26:04am
tic-toc.cpp:            30/ 9/2008  2:57:15pm